Remove all lint errors and warnings
