+++ src/ui/img/icon_sawtooth.tsx
@@ -1,0 +1,6 @@
+export const svgSawtooth = <svg className="svg-inline--fa" width="30" height="30">
+  <svg viewBox="0 0 20 20">
+    <polyline style={{ fill: 'none', stroke: 'black', strokeWidth: '2.5px', strokeLinecap: 'round', strokeMiterlimit: 7.46 }}
+      points="0 9.847 6.379 3.135 6.269 16.048 14.035 3.611 13.973 16.424 19.5 9.348" />
+  </svg></svg>
+
